Check off more TODOs Try and make the effects list more accurate Fix queue not respecting firt selection

+++ app/src/main/java/org/helllabs/android/xmp/Xmp.kt
@@ -118,8 +118,8 @@ object Xmp {
         row: Int,
         rowNotes: ByteArray,
         rowInstruments: ByteArray,
-        rowFxType: IntArray,
-        rowFxParm: IntArray
+        rowFxType: ByteArray,
+        rowFxParm: ByteArray
     )
 
     external fun getSampleData(
